+++ src/main/java/org/betonquest/betonquest/api/instruction/argument/types/IdentifierParser.java
@@ -1,8 +1,8 @@
 package org.betonquest.betonquest.api.instruction.argument.types;
 
 import org.betonquest.betonquest.api.config.quest.QuestPackage;
+import org.betonquest.betonquest.api.identifier.Identifier;
 import org.betonquest.betonquest.api.instruction.argument.PackageArgument;
-import org.betonquest.betonquest.api.quest.QuestException;
 
 /**
  * Parses a string to an identifier.
@@ -15,10 +15,10 @@ public class IdentifierParser implements PackageArgument<String> {
     }
 
     @Override
-    public String apply(final QuestPackage pack, final String string) throws QuestException {
-        if (string.contains(".")) {
+    public String apply(final QuestPackage pack, final String string) {
+        if (string.contains(Identifier.SEPARATOR)) {
             return string;
         }
-        return pack.getQuestPath() + "." + string;
+        return pack.getQuestPath() + Identifier.SEPARATOR + string;
     }
 }