+++ src/main/java/org/betonquest/betonquest/quest/condition/conversation/InConversationConditionFactory.java
@@ -5,7 +5,7 @@ import org.betonquest.betonquest.api.instruction.variable.Variable;
 import org.betonquest.betonquest.api.quest.QuestException;
 import org.betonquest.betonquest.api.quest.condition.PlayerCondition;
 import org.betonquest.betonquest.api.quest.condition.PlayerConditionFactory;
-import org.betonquest.betonquest.id.ConversationID;
+import org.betonquest.betonquest.conversation.ConversationID;
 
 /**
  * Factory for {@link InConversationCondition}s.