+++ CHANGELOG.md
@@ -88,7 +88,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `menu` Conversation IO displaying and scroll behavior rework
 - `objectives` event listener are now registered on reload instead when the first player gets it active
 - `npcmove` event - remove strange code effecting the end of a navigation to may return to a previous location in some cases
-- `npc_conversations` show the combat message only if there is a bound conversation  
+- `npc_conversations` show the combat message only if there is a bound conversation
+- Addressing cross Packages now use `>` instead of `.` to solve conflicts in Variables
 ### Deprecated
 ### Removed
 - undocumented prefix feature in conversation
@@ -98,6 +99,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `variables` section in favor for `constants`
 - `MMOItems` conditions, events and `mmoitemcraft` objective in favor of the built-in ones
 - `journal.chars_per_page` was deleted from the config.yml 
+- undocumented way of addressing packages in some variables
 ### Fixed
 - Reloading plugin did not reload Menu config
 - potions generated not extended/upgraded since 2.2.1